+++ atividade4_CaioLuizbatisti.py
@@ -6,7 +6,7 @@
 #Escreva um programa que leia 3 dados de entrada: nome da pessoa, idade e sexo e informe se a pessoa será aceita ou não para o serviço.
 #obs:para o sexo deve ser lido apenas um caractere que pode ser “F” ou “f” “m” ou “M” para masculino e feminino, qualquer coisa diferente deve ser informado
 nome = input("digite o nome:")
-idade =int(input("digite a idade:"))
+idade =float(input("digite a idade:"))
 genero = input("digite seu genero:")
 if genero == ("M") or ("m") and idade >= 18 and idade < 39:
     print("voce pode servir.")
